+++ query/statement.go
@@ -1,0 +1,39 @@
+package query
+
+type SelectStmt struct {
+	AllFields  bool
+	FieldNames []string
+	Fields     []Expression
+	Where      *WhereStmt
+	Order      *OrderStmt
+	Limit      *LimitStmt
+	GroupBy    *GroupByStmt
+}
+
+type WhereStmt struct {
+	Expr Expression
+}
+
+type OrderField struct {
+	Name  string
+	Field Expression
+	Order TokenType
+}
+
+type OrderStmt struct {
+	Orders []OrderField
+}
+
+type GroupByField struct {
+	Name string
+	Expr Expression
+}
+
+type GroupByStmt struct {
+	Fields []GroupByField
+}
+
+type LimitStmt struct {
+	Start int
+	Count int
+}
